extract getUserRow
