+++ static/js/src/advantage/users/AccountUsers.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import { Users, OrganisationName } from "./types";
 import Organisation from "./components/Organisation";
-import TableView from "./components/TableView";
+import TableView from "./components/TableView/TableView";
 
 const AccountUsers: React.FC<{
   organisationName: OrganisationName;
