+++ static/js/src/advantage/users/components/TableView/TableView.tsx
@@ -1,0 +1,58 @@
+import React, { useState } from "react";
+import MainTable from "@canonical/react-components/dist/components/MainTable";
+
+import { Users } from "../../types";
+import { getUserRow } from "./components";
+
+export type UserRowVariant = "regular" | "editing" | "disabled";
+
+type UserInEditMode = string | null;
+type UserId = string;
+
+const getVariant = (userId: UserId, userInEditMode: UserInEditMode) => {
+  if (userInEditMode === null) {
+    return "regular";
+  } else if (userId !== userInEditMode) {
+    return "disabled";
+  } else {
+    return "editing";
+  }
+};
+
+const TableView: React.FC<{ users: Users }> = ({ users }) => {
+  const [userInEditMode, setUserInEditMode] = useState<string | null>(null);
+  const dismissEditMode = () => setUserInEditMode(null);
+
+  return (
+    <MainTable
+      responsive
+      headers={[
+        {
+          content: "email",
+        },
+        {
+          content: "role",
+        },
+        {
+          content: "date added",
+        },
+        {
+          content: "last sign in",
+        },
+        {
+          content: "actions",
+        },
+      ]}
+      rows={users.map((user) =>
+        getUserRow({
+          user,
+          variant: getVariant(user.id, userInEditMode),
+          setUserInEditMode,
+          dismissEditMode,
+        })
+      )}
+    />
+  );
+};
+
+export default TableView;
